Move test logger to own file
